refactor(python-sdk): split envd RPC plumbing into transport/interceptors/client modules
e2b/envd/transport.py keeps only the shared pyqwest transports and proxy
handling; the interceptors move to e2b/envd/interceptors.py and the codec,
stream casts, and client factories to e2b/envd/client.py. The sync flag on
create_rpc_client is replaced by separate create_rpc_client and
create_async_rpc_client functions.
