+++ packages/python-sdk/e2b/sandbox_sync/commands/pty.py
@@ -24,7 +24,7 @@ from e2b.envd.rpc import (
     handle_rpc_exception_with_health,
     timeout_to_ms,
 )
-from e2b.envd.transport import as_stream, create_rpc_client
+from e2b.envd.client import as_stream, create_rpc_client
 from e2b.sandbox.commands.command_handle import PtySize
 from e2b.sandbox_sync.commands.command_handle import CommandHandle
 
@@ -59,7 +59,6 @@ class Pty:
             process_connect.ProcessClientSync,
             self._envd_api_url,
             self._connection_config,
-            sync=True,
         )
 
     @property
